s390: vfio-ap: sysfs interface to view guest matrix
Provides a sysfs interface to view the AP matrix configured for the
guest that is using the mdev matrix device.

The relevant sysfs structures are:

/sys/devices/vfio_ap
... [matrix]
...... [mdev_supported_types]
......... [vfio_ap-passthrough]
............ [devices]
...............[$uuid]
.................. guest_matrix

To view the matrix configured for the guest,
print the guest_matrix file:

cat guest_matrix

If no guest is using the device, then the output will be an emtpy
string.

Signed-off-by: Tony Krowiak <akrowiak@linux.ibm.com>