crypto: zip: make several functions static
The functions zip_ctx_exit, zip_compress, zip_decompress and zip_ctx_init
are local to the source and do not need to be in global scope, so make
them static.

Cleans up sparse warnings:
symbol 'zip_ctx_exit' was not declared. Should it be static?
symbol 'zip_compress' was not declared. Should it be static?
symbol 'zip_decompress' was not declared. Should it be static?
symbol 'zip_ctx_init' was not declared. Should it be static?

Signed-off-by: Colin Ian King <colin.king@canonical.com>