usb/host/pci-quirks: Only reset USB bus on NVIDIA devices
Currently, on the AMD board Asus F2A85-M Pro there is a 100 ms delay as
the USB bus of each of the two OHCI PCI devices is reset. As a 50 ms
delay is done per the USB specification.

Commit c6187597 (OHCI: final fix for NVIDIA problems (I hope))
unconditionally does the bus reset for
 all chipsets, while it was only doen for NVIDIA chipsets before.

As it should not be needed for non-NVIDIA chipsets, only do the reset
for Nvidia devices.

Tested on Asus F2A85-M PRO and ASRock E350M1. The USB keyboard works and
the LUKS passphrase can be e
ntered.

Signed-off-by: Paul Menzel <pmenzel@molgen.mpg.de>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: linux-usb@vger.kernel.org
Cc: Alan Stern <stern@rowland.harvard.edu>
Cc: linux-kernel@vger.kernel.org
Cc: stable@vger.kernel.org