drm/msm: Disable mdp5 crtc when there are no active planes
Unlike other compositors, we don't get a crtc disable from weston
when the cable is unplugged. As such, when the cable is re-plugged
the kernel doesn't detect an enable/mode change and initiates a
simple plane update instead of a modeset.

This patch clears the mode when all planes are off.

Signed-off-by: Sean Paul <seanpaul@chromium.org>